+++ src/sora_signaling.cpp
@@ -138,6 +138,9 @@ void SoraSignaling::Redirect(std::string url) {
         return;
       }
 
+      // Disconnect と重複しないように Redirecting の場合のみ発火させる
+      self->SendSelfOnWsClose(ec);
+
       // close 処理に成功してても失敗してても処理は続ける
       if (ec) {
         RTC_LOG(LS_WARNING) << "Redirect error: ec=" << ec.message();
@@ -171,13 +174,14 @@ void SoraSignaling::Redirect(std::string url) {
           ws.reset(
               new Websocket(Websocket::ssl_tag(), *self->config_.io_context,
                             self->config_.insecure, self->config_.client_cert,
-                            self->config_.client_key));
+                            self->config_.client_key, self->config_.ca_cert));
         } else {
           ws.reset(new Websocket(
               Websocket::https_proxy_tag(), *self->config_.io_context,
               self->config_.insecure, self->config_.client_cert,
-              self->config_.client_key, self->config_.proxy_url,
-              self->config_.proxy_username, self->config_.proxy_password));
+              self->config_.client_key, self->config_.ca_cert,
+              self->config_.proxy_url, self->config_.proxy_username,
+              self->config_.proxy_password));
         }
       } else {
         ws.reset(new Websocket(*self->config_.io_context));
@@ -412,10 +416,10 @@ void SoraSignaling::DoSendConnect(bool redirect) {
     m["forwarding_filter"] = obj;
   }
 
-  RTC_LOG(LS_INFO) << "Send type=connect: " << boost::json::serialize(m);
-  ws_->WriteText(
-      boost::json::serialize(m),
-      [self = shared_from_this()](boost::system::error_code, size_t) {});
+  std::string text = boost::json::serialize(m);
+  RTC_LOG(LS_INFO) << "Send type=connect: " << text;
+  WsWriteSignaling(std::move(text), [self = shared_from_this()](
+                                        boost::system::error_code, size_t) {});
 }
 
 void SoraSignaling::DoSendPong() {
@@ -441,12 +445,15 @@ void SoraSignaling::DoSendPong(
 
 void SoraSignaling::DoSendUpdate(const std::string& sdp, std::string type) {
   boost::json::value m = {{"type", type}, {"sdp", sdp}};
+  std::string text = boost::json::serialize(m);
   if (dc_ && using_datachannel_ && dc_->IsOpen("signaling")) {
     // DataChannel が使える場合は DataChannel に送る
-    SendDataChannel("signaling", boost::json::serialize(m));
+    SendDataChannel("signaling", text);
+    SendOnSignalingMessage(SoraSignalingType::DATACHANNEL,
+                           SoraSignalingDirection::SENT, std::move(text));
   } else if (ws_) {
-    ws_->WriteText(
-        boost::json::serialize(m),
+    WsWriteSignaling(
+        std::move(text),
         [self = shared_from_this()](boost::system::error_code, size_t) {});
   }
 }
@@ -510,7 +517,7 @@ SoraSignaling::CreatePeerConnection(boost::json::value jconfig) {
   //
   // それを解消するために tls_cert_verifier を設定して自前で検証を行う。
   dependencies.tls_cert_verifier = std::unique_ptr<rtc::SSLCertificateVerifier>(
-      new RTCSSLVerifier(config_.insecure));
+      new RTCSSLVerifier(config_.insecure, config_.ca_cert));
 
   // Proxy を設定する
   if (!config_.proxy_url.empty() && config_.network_manager != nullptr &&
@@ -539,7 +546,8 @@ SoraSignaling::CreatePeerConnection(boost::json::value jconfig) {
       pi.username = config_.proxy_username;
     }
     if (!config_.proxy_password.empty()) {
-      pi.password = rtc::revive::CryptString(RawCryptString(config_.proxy_password));
+      pi.password =
+          rtc::revive::CryptString(RawCryptString(config_.proxy_password));
     }
     std::string proxy_agent = "Sora C++ SDK";
     if (!config_.proxy_agent.empty()) {
@@ -590,6 +598,7 @@ void SoraSignaling::DoInternalDisconnect(
       boost::system::error_code tec;
       self->closing_timeout_timer_.cancel(tec);
       auto ws_reason = self->ws_->reason();
+      self->SendOnWsClose(ws_reason);
       std::string message =
           "Unintended disconnect WebSocket during Disconnect process: ec=" +
           ec.message() + " wscode=" + std::to_string(ws_reason.code) +
@@ -599,11 +608,11 @@ void SoraSignaling::DoInternalDisconnect(
   }
 
   if (using_datachannel_ && ws_connected_) {
-    webrtc::DataBuffer disconnect = ConvertToDataBuffer(
-        "signaling",
+    std::string text =
         force_error_code == boost::none
             ? R"({"type":"disconnect","reason":"NO-ERROR"})"
-            : R"({"type":"disconnect","reason":")" + reason + "\"}");
+            : R"({"type":"disconnect","reason":")" + reason + "\"}";
+    webrtc::DataBuffer disconnect = ConvertToDataBuffer("signaling", text);
     dc_->Close(
         disconnect,
         [self = shared_from_this(), on_close, force_error_code,
@@ -623,6 +632,7 @@ void SoraSignaling::DoInternalDisconnect(
             boost::system::error_code tec;
             self->closing_timeout_timer_.cancel(tec);
             auto ws_reason = self->ws_->reason();
+            self->SendOnWsClose(ws_reason);
             std::string ws_reason_str =
                 " wscode=" + std::to_string(ws_reason.code) +
                 " wsreason=" + ws_reason.reason.c_str();
@@ -652,9 +662,12 @@ void SoraSignaling::DoInternalDisconnect(
           };
         },
         config_.disconnect_wait_timeout);
+
+    SendOnSignalingMessage(SoraSignalingType::DATACHANNEL,
+                           SoraSignalingDirection::SENT, std::move(text));
   } else if (using_datachannel_ && !ws_connected_) {
-    webrtc::DataBuffer disconnect = ConvertToDataBuffer(
-        "signaling", R"({"type":"disconnect","reason":"NO-ERROR"})");
+    std::string text = R"({"type":"disconnect","reason":"NO-ERROR"})";
+    webrtc::DataBuffer disconnect = ConvertToDataBuffer("signaling", text);
     dc_->Close(
         disconnect,
         [on_close](boost::system::error_code ec) {
@@ -667,10 +680,13 @@ void SoraSignaling::DoInternalDisconnect(
                    "Succeeded to close DataChannel");
         },
         config_.disconnect_wait_timeout);
+
+    SendOnSignalingMessage(SoraSignalingType::DATACHANNEL,
+                           SoraSignalingDirection::SENT, std::move(text));
   } else if (!using_datachannel_ && ws_connected_) {
     boost::json::value disconnect = {{"type", "disconnect"},
                                      {"reason", "NO-ERROR"}};
-    ws_->WriteText(
+    WsWriteSignaling(
         boost::json::serialize(disconnect),
         [self = shared_from_this(), on_close](boost::system::error_code ec,
                                               std::size_t) {
@@ -695,9 +711,10 @@ void SoraSignaling::DoInternalDisconnect(
           self->on_ws_close_ = [self, on_close](boost::system::error_code ec) {
             boost::system::error_code tec;
             self->closing_timeout_timer_.cancel(tec);
+            auto reason = self->ws_->reason();
+            self->SendOnWsClose(reason);
             bool ec_error = ec != boost::beast::websocket::error::closed;
             if (ec_error) {
-              auto reason = self->ws_->reason();
               on_close(false, SoraSignalingErrorCode::CLOSE_FAILED,
                        "Failed to close WebSocket: ec=" + ec.message() +
                            " wscode=" + std::to_string(reason.code) +
@@ -809,6 +826,7 @@ void SoraSignaling::OnRead(boost::system::error_code ec,
                             ? SoraSignalingErrorCode::WEBSOCKET_ONCLOSE
                             : SoraSignalingErrorCode::WEBSOCKET_ONERROR;
       auto reason = ws_->reason();
+      SendOnWsClose(reason);
       std::string reason_str = " wscode=" + std::to_string(reason.code) +
                                " wsreason=" + reason.reason.c_str();
       SendOnDisconnect(error_code, "Failed to read WebSocket: ec=" +
@@ -821,15 +839,16 @@ void SoraSignaling::OnRead(boost::system::error_code ec,
     }
     if (state_ == State::Connected && using_datachannel_ && ws_connected_) {
       // DataChanel で reason: "WEBSOCKET-ONCLOSE" または "WEBSOCKET-ONERROR" を送る事を試みてから終了する
-      webrtc::DataBuffer disconnect = ConvertToDataBuffer(
-          "signaling",
+      std::string text =
           ec == boost::beast::websocket::error::closed
               ? R"({"type":"disconnect","reason":"WEBSOCKET-ONCLOSE"})"
-              : R"({"type":"disconnect","reason":"WEBSOCKET-ONERROR"})");
+              : R"({"type":"disconnect","reason":"WEBSOCKET-ONERROR"})";
+      webrtc::DataBuffer disconnect = ConvertToDataBuffer("signaling", text);
       auto error_code = ec == boost::beast::websocket::error::closed
                             ? SoraSignalingErrorCode::WEBSOCKET_ONCLOSE
                             : SoraSignalingErrorCode::WEBSOCKET_ONERROR;
       auto reason = ws_->reason();
+      SendOnWsClose(reason);
       std::string reason_str = " wscode=" + std::to_string(reason.code) +
                                " wsreason=" + reason.reason.c_str();
       state_ = State::Closing;
@@ -851,6 +870,9 @@ void SoraSignaling::OnRead(boost::system::error_code ec,
                     ec.message() + reason_str);
           },
           config_.disconnect_wait_timeout);
+
+      SendOnSignalingMessage(SoraSignalingType::DATACHANNEL,
+                             SoraSignalingDirection::SENT, std::move(text));
       return;
     }
 
@@ -877,11 +899,18 @@ void SoraSignaling::OnRead(boost::system::error_code ec,
   }
 
   if (type == "redirect") {
+    SendOnSignalingMessage(SoraSignalingType::WEBSOCKET,
+                           SoraSignalingDirection::RECEIVED, std::move(text));
+
     const std::string location = m.at("location").as_string().c_str();
     Redirect(location);
     // Redirect の中で次の Read をしているのでここで return する
     return;
   } else if (type == "offer") {
+    // 後続で使っているので text はコピーする
+    SendOnSignalingMessage(SoraSignalingType::WEBSOCKET,
+                           SoraSignalingDirection::RECEIVED, text);
+
     const std::string sdp = m.at("sdp").as_string().c_str();
 
     std::string video_mid;
@@ -1022,7 +1051,7 @@ void SoraSignaling::OnRead(boost::system::error_code ec,
                     }
 
                     boost::json::value m = {{"type", "answer"}, {"sdp", sdp}};
-                    self->ws_->WriteText(
+                    self->WsWriteSignaling(
                         boost::json::serialize(m),
                         [self](boost::system::error_code, size_t) {});
                   });
@@ -1036,6 +1065,10 @@ void SoraSignaling::OnRead(boost::system::error_code ec,
     if (pc_ == nullptr) {
       return;
     }
+
+    SendOnSignalingMessage(SoraSignalingType::WEBSOCKET,
+                           SoraSignalingDirection::RECEIVED, std::move(text));
+
     std::string answer_type = type == "update" ? "update" : "re-answer";
     const std::string sdp = m.at("sdp").as_string().c_str();
     if (!CheckSdp(sdp)) {
@@ -1118,8 +1151,11 @@ void SoraSignaling::OnRead(boost::system::error_code ec,
     auto it = m.as_object().find("ignore_disconnect_websocket");
     if (it != m.as_object().end() && it->value().as_bool() && ws_connected_) {
       RTC_LOG(LS_INFO) << "Close WebSocket for DataChannel";
-      ws_->Close([self = shared_from_this()](boost::system::error_code) {},
-                 config_.websocket_close_timeout);
+      ws_->Close(
+          [self = shared_from_this()](boost::system::error_code ec) {
+            self->SendSelfOnWsClose(ec);
+          },
+          config_.websocket_close_timeout);
       ws_connected_ = false;
 
       return;
@@ -1172,12 +1208,12 @@ void SoraSignaling::DoConnect() {
       if (config_.proxy_url.empty()) {
         ws.reset(new Websocket(Websocket::ssl_tag(), *config_.io_context,
                                config_.insecure, config_.client_cert,
-                               config_.client_key));
+                               config_.client_key, config_.ca_cert));
       } else {
         ws.reset(new Websocket(
             Websocket::https_proxy_tag(), *config_.io_context, config_.insecure,
-            config_.client_cert, config_.client_key, config_.proxy_url,
-            config_.proxy_username, config_.proxy_password));
+            config_.client_cert, config_.client_key, config_.ca_cert,
+            config_.proxy_url, config_.proxy_username, config_.proxy_password));
       }
     } else {
       ws.reset(new Websocket(*config_.io_context));
@@ -1303,6 +1339,14 @@ void SoraSignaling::ResetEncodingParameters() {
   sender->SetParameters(parameters);
 }
 
+void SoraSignaling::WsWriteSignaling(std::string text,
+                                     Websocket::write_callback_t on_write) {
+  ws_->WriteText(text, on_write);
+
+  SendOnSignalingMessage(SoraSignalingType::WEBSOCKET,
+                         SoraSignalingDirection::SENT, std::move(text));
+}
+
 void SoraSignaling::SendOnDisconnect(SoraSignalingErrorCode ec,
                                      std::string message) {
   if (ec != SoraSignalingErrorCode::CLOSE_SUCCEEDED) {
@@ -1316,6 +1360,28 @@ void SoraSignaling::SendOnDisconnect(SoraSignalingErrorCode ec,
       ob->OnDisconnect(ec, std::move(message));
     }
   });
+}
+
+void SoraSignaling::SendOnSignalingMessage(SoraSignalingType type,
+                                           SoraSignalingDirection direction,
+                                           std::string message) {
+  if (auto ob = config_.observer.lock(); ob) {
+    ob->OnSignalingMessage(type, direction, std::move(message));
+  }
+}
+
+void SoraSignaling::SendOnWsClose(
+    const boost::beast::websocket::close_reason& reason) {
+  if (auto ob = config_.observer.lock(); ob) {
+    ob->OnWsClose(reason.code, reason.reason.c_str());
+  }
+}
+
+void SoraSignaling::SendSelfOnWsClose(boost::system::error_code ec) {
+  auto close_reason = boost::beast::websocket::close_reason(
+      ec ? 4999 : boost::beast::websocket::close_code::normal);
+  close_reason.reason = "SELF-CLOSED";
+  SendOnWsClose(close_reason);
 }
 
 webrtc::DataBuffer SoraSignaling::ConvertToDataBuffer(
@@ -1437,8 +1503,8 @@ void SoraSignaling::OnIceCandidate(
           return;
         }
 
-        self->ws_->WriteText(boost::json::serialize(m),
-                             [self](boost::system::error_code, size_t) {});
+        self->WsWriteSignaling(boost::json::serialize(m),
+                               [self](boost::system::error_code, size_t) {});
       });
 }
 
@@ -1533,7 +1599,7 @@ void SoraSignaling::OnMessage(
     return;
   }
 
-  boost::json::error_code ec;
+  boost::system::error_code ec;
   auto json = boost::json::parse(data, ec);
   if (ec) {
     RTC_LOG(LS_ERROR) << "JSON Parse Error ec=" << ec.message();
@@ -1541,6 +1607,9 @@ void SoraSignaling::OnMessage(
   }
 
   if (label == "signaling") {
+    SendOnSignalingMessage(SoraSignalingType::DATACHANNEL,
+                           SoraSignalingDirection::RECEIVED, std::move(data));
+
     const std::string type = json.at("type").as_string().c_str();
     if (type == "re-offer") {
       const std::string sdp = json.at("sdp").as_string().c_str();